requirements.txt
blessed~=1.22.0
google-genai
dotenv